add and tidy util functions
