+++ src/rendering/renderers/shared/buffer/Buffer.ts
@@ -1,5 +1,5 @@
 import EventEmitter from 'eventemitter3';
-import { generateUID } from '../texture/utils/generateUID';
+import { uid } from '../../../../utils/data/uid';
 
 import type { BindResource } from '../../gpu/shader/BindResource';
 import type { BufferUsage } from './const';
@@ -33,7 +33,7 @@ export class Buffer extends EventEmitter<{
 }> implements BindResource
 {
     public readonly resourceType = 'buffer';
-    public resourceId = generateUID();
+    public resourceId = uid();
 
     public touched = 0;
     public readonly uid = UID++;
@@ -86,7 +86,7 @@ export class Buffer extends EventEmitter<{
             if (oldData.length !== value.length)
             {
                 this.descriptor.size = value.byteLength;
-                this.resourceId = generateUID();
+                this.resourceId = uid();
 
                 this.emit('change', this);
             }
